+ Login and password authorization.
On the "Settings" page, when loading the View, a list is displayed from the "Status" and "Predmet" tables.
Added the ability to delete and add positions, as well as dynamic page refresh.

On the "Persons" page, when loading the View, a list from the "Status" table is displayed.

-
JournalViev;

in procces:
add person